+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Fixed a performance issue causing slowdowns in tokenizing (using the default `what = "word"`) corpora with large numbers of documents that contain social media tags and URLs that needed to be preserved (such a large corpus of Tweets).
 * Updated the (default) "word" tokenizer to preserve hashtags and usernames better with non-ASCII text, and made these patterns user-configurable in `quanteda_options()`.  The following are now preserved: `#政治` as well as Weibo-style hashtags such as `#英国首相#".
+* Added new methods `char_select()`, `char_keep()`, and `char_remove()` for easy manipulation of character vectors.
 
 
 # quanteda 2.0.1